Implement ctrl-selection with keyboard: bug #716663
Holding the control key allows to move around with the arrow keys and
use the space bar to select photos. The cursor is shown with a
selection-width border with the same gray color used for normal borders.

As an added benefit the same cursor border is used when control-clicking
a selection away with the mouse.